[secrets][bare-expo] Fixes to auto-gen & copy of GoogleServices-Info.plist

+++ apps/bare-expo/ios/Build-Phases/cp-bundle-resources-conditionally.sh
@@ -5,5 +5,7 @@ set -exo pipefail
 SRC_FILE="${PROJECT_DIR}/BareExpo/GoogleService-Info.plist"
 DST_FILE="${BUILT_PRODUCTS_DIR}/${PRODUCT_NAME}.app/GoogleService-Info.plist"
 if [ -f $SRC_FILE ]; then
-    cp -r "$SRC_FILE" "$DST_FILE"
+    if [[ $(grep -L "NO_CLIENT_ID" "$SRC_FILE") ]]; then
+        cp -r "$SRC_FILE" "$DST_FILE"
+    fi
 fi
